requirements.txt
Flask==2.3.3
gunicorn==21.2.0
opencv-python-headless==4.10.0.84
roboflow==1.1.39
cohere==5.13.4
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
google-genai==0.1.0
Pillow==10.3.0
python-dotenv==1.0.1